pyproject.toml
[project]
name = "signature-durability-benchmark"
version = "0.1.0"
requires-python = "==3.12.*"
dependencies = [
    "jsonschema>=4.25,<5",
    "matplotlib>=3.10,<4",
    "numpy>=2.3,<3",
    "pandas>=2.3,<3",
    "pyyaml>=6,<7",
    "scikit-learn>=1.7,<2",
    "scipy>=1.15,<2",
]

[project.optional-dependencies]
dev = ["pytest>=8.4,<9"]

[project.scripts]
signature-durability-benchmark = "signature_durability_benchmark.cli:main"

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
